refactor courtesy of @JoshVanL
