+++ tests/integration/framework/process/grpc/app/options.go
@@ -19,6 +19,8 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 
+	testpb "github.com/dapr/dapr/tests/integration/framework/process/grpc/app/proto"
+
 	commonv1 "github.com/dapr/dapr/pkg/proto/common/v1"
 	rtv1 "github.com/dapr/dapr/pkg/proto/runtime/v1"
 	procgrpc "github.com/dapr/dapr/tests/integration/framework/process/grpc"
@@ -37,6 +39,7 @@ type options struct {
 	listInputBindFn    func(context.Context, *emptypb.Empty) (*rtv1.ListInputBindingsResponse, error)
 	onBindingEventFn   func(context.Context, *rtv1.BindingEventRequest) (*rtv1.BindingEventResponse, error)
 	healthCheckFn      func(context.Context, *emptypb.Empty) (*rtv1.HealthCheckResponse, error)
+	pingFn             func(context.Context, *testpb.PingRequest) (*testpb.PingResponse, error)
 }
 
 func WithGRPCOptions(opts ...procgrpc.Option) func(*options) {
@@ -90,6 +93,12 @@ func WithOnBindingEventFn(fn func(context.Context, *rtv1.BindingEventRequest) (*
 func WithHealthCheckFn(fn func(context.Context, *emptypb.Empty) (*rtv1.HealthCheckResponse, error)) func(*options) {
 	return func(opts *options) {
 		opts.healthCheckFn = fn
+	}
+}
+
+func WithPingFn(fn func(context.Context, *testpb.PingRequest) (*testpb.PingResponse, error)) func(*options) {
+	return func(opts *options) {
+		opts.pingFn = fn
 	}
 }
