+++ tests/integration/framework/process/grpc/app/proto/test.v1.proto
